requirements.txt
matplotlib>=1.5.1
numpy>=1.13.1
pandas>=0.18.1
Pillow>=2.8.2
scipy>=0.18.1
seaborn>=0.8.1
